fix: 修正法人資料單位、DB過濾、scheduler重複寫入
- twse.py: fetch_institutional 存法人買超由「股」改為「張」(÷1000)
- screener.py: calc_chip_ratios capital=0時仍回傳淨買超張數供顯示
- scheduler.py: _already_fetched 只跳過 success；_log_fetch ON CONFLICT DO UPDATE；job1/job2 pg_insert ON CONFLICT DO NOTHING；job4 先刪今日再插入
- init_data.py: step3 先刪今日 screening_result 再插入；step2b insert 前過濾非stock_list代號
- 新增 price.py (yfinance 批次下載) 和 update_capital.py script
- Dashboard.tsx: job狀態列改顯示中文標籤
- README/plan.md: 修正欄位名、評分表、加入資料範圍限制說明

+++ backend/app/db/models.py
@@ -1,5 +1,5 @@
 from datetime import date, datetime
-from sqlalchemy import String, Integer, Float, Boolean, Date, DateTime, Text, UniqueConstraint
+from sqlalchemy import String, Integer, BigInteger, Float, Boolean, Date, DateTime, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.base import Base
 
@@ -47,10 +47,10 @@ class MarginTrading(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     code: Mapped[str] = mapped_column(String(10), index=True)
     trade_date: Mapped[date] = mapped_column(Date, index=True)
-    margin_balance: Mapped[int] = mapped_column(Integer, default=0)
-    margin_change: Mapped[int] = mapped_column(Integer, default=0)
-    short_balance: Mapped[int] = mapped_column(Integer, default=0)
-    short_change: Mapped[int] = mapped_column(Integer, default=0)
+    margin_balance: Mapped[int] = mapped_column(BigInteger, default=0)
+    margin_change: Mapped[int] = mapped_column(BigInteger, default=0)
+    short_balance: Mapped[int] = mapped_column(BigInteger, default=0)
+    short_change: Mapped[int] = mapped_column(BigInteger, default=0)
 
 
 class Shareholding(Base):
